requirements.txt
advertools==0.13.2
alembic==1.8.1
anaconda-navigator==2.1.1
appdirs==1.4.4
argh==0.26.2
asgiref==3.6.0
atomicwrites==1.4.0
audioread==2.1.9
Automat==20.2.0
backports.weakref==1.0.post1
bkcharts==0.2
black==23.1.0
bootstrap4==0.1.0
boto==2.49.0
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.3.1
certifi==2021.10.8
click==8.1.3
cloudscraper==1.2.71
clyent==1.2.2
conda==4.12.0
conda-build==3.21.5
conda-verify==3.4.2
constantly==15.1.0
crispy-bootstrap4==2022.1
cssselect==1.1.0
cycler==0.10.0
cytoolz==0.11.0
daal4py==2021.3.0
dask==2021.10.0
distlib==0.3.4
Django==4.0.2
django-allauth==0.13.0
django-crispy-forms==2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
dtw==1.4.0
entrypoints==0.3
et-xmlfile==1.1.0
fake-headers==1.0.2
fastdtw==0.3.4
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
fonttools==4.25.0
gmpy2==2.0.8
google-api-core==2.11.1
google-auth==2.22.0
google-cloud==0.34.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.13.0
hyperlink==21.0.0
incremental==21.3.0
inflection==0.5.1
itemadapter==0.5.0
itemloaders==1.0.4
jmespath==1.0.0
librosa==0.9.1
llvmlite==0.37.0
locket==0.2.1
Mako==1.2.1
mccabe==0.6.1
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.30
navigator-updater==0.2.1
nltk==3.6.5
oauthlib==3.2.2
opencv-python==4.5.5.64
openlocationcode==1.0.1
outcome==1.1.0
packaging==23.0
pandas==1.3.4
parsel==1.6.0
pathspec==0.11.0
patsy==0.5.2
pep8==1.7.1
Pillow==8.4.0
pkginfo==1.7.1
platformdirs==3.1.1
ply==3.11
pooch==1.6.0
Protego==0.2.1
proto-plus==1.22.3
protobuf==4.24.2
psycopg2-binary==2.9.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycosat==0.6.3
pycurl==7.44.1
PyDispatcher==2.0.5
pyls-spyder==0.4.0
pymongo==3.12.3
pyodbc===4.0.0-unsupported
PyPDF2==3.0.1
pytest==6.2.4
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
queuelib==1.6.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
resampy==0.2.2
rsa==4.9
scikit-image==0.18.3
scikit-learn-intelex==2021.20210714.100439
Scrapy==2.6.1
scrapy-splash==0.8.0
selenium==4.1.3
service-identity==21.1.0
simplegeneric==0.8.1
sip==4.19.13
sklearn==0.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SoundFile==0.10.3.post1
Sphinx==4.2.0
sqlparse==0.2.4
TBB==0.2
terminado==0.9.4
tldextract==3.2.1
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.4.0
twython==3.9.1
typing_extensions==4.5.0
unicodecsv==0.14.1
urllib3==1.26.7
virtualenv==20.13.1
w3lib==1.22.0
webdriver-manager==3.5.4
webencodings==0.5.1
wsproto==1.1.0
xlwings==0.24.9
xlwt==1.3.0
zict==2.0.0
zope.event==4.5.0
